Add `=>` expression support to test_matrix macro
Required that `TestCaseExpression` and all it's descendant structs be
`Clone` so that TestMatrix can clone them for each TestCase it creates.
Also made TestComment Clone for parity.

Implemented quote::ToTokens for TestComment so that the "illegal
comment" error generated by TestMatrix can point to the correct span of
where the illegal comment begins. On stable Rust, this just highlights
the first token, the semicolon, but on nightly Rust (which has
`quote::Span::join`) it will highlight the entire illegal comment.